+++ src/Persistence/src/Dao/SQLite/SQLiteFlightDao.cpp
@@ -78,6 +78,7 @@ bool SQLiteFlightDao::add(Flight &flight) noexcept
         "  description,"
         "  user_aircraft_seq_nr,"
         "  surface_type,"
+        "  surface_condition,"
         "  ground_altitude,"
         "  ambient_temperature,"
         "  total_air_temperature,"
@@ -88,6 +89,8 @@ bool SQLiteFlightDao::add(Flight &flight) noexcept
         "  pitot_icing,"
         "  structural_icing,"
         "  precipitation_state,"
+        "  on_any_runway,"
+        "  in_parking_state,"
         "  in_clouds,"
         "  start_local_sim_time,"
         "  start_zulu_sim_time,"
@@ -99,6 +102,7 @@ bool SQLiteFlightDao::add(Flight &flight) noexcept
         " :description,"
         " :user_aircraft_seq_nr,"
         " :surface_type,"
+        " :surface_condition,"
         " :ground_altitude,"
         " :ambient_temperature,"
         " :total_air_temperature,"
@@ -109,6 +113,8 @@ bool SQLiteFlightDao::add(Flight &flight) noexcept
         " :pitot_icing,"
         " :structural_icing,"
         " :precipitation_state,"
+        " :on_any_runway,"
+        " :in_parking_state,"
         " :in_clouds,"
         " :start_local_sim_time,"
         " :start_zulu_sim_time,"
@@ -124,6 +130,7 @@ bool SQLiteFlightDao::add(Flight &flight) noexcept
     // Sequence number starts at 1
     query.bindValue(":user_aircraft_seq_nr", flight.getUserAircraftIndex() + 1);
     query.bindValue(":surface_type", Enum::underly(flightCondition.surfaceType));
+    query.bindValue(":surface_condition", Enum::underly(flightCondition.surfaceCondition));
     query.bindValue(":ground_altitude", flightCondition.groundAltitude);
     query.bindValue(":ambient_temperature", flightCondition.ambientTemperature);
     query.bindValue(":total_air_temperature", flightCondition.totalAirTemperature);
@@ -134,6 +141,8 @@ bool SQLiteFlightDao::add(Flight &flight) noexcept
     query.bindValue(":pitot_icing", flightCondition.pitotIcingPercent);
     query.bindValue(":structural_icing", flightCondition.structuralIcingPercent);
     query.bindValue(":precipitation_state", Enum::underly(flightCondition.precipitationState));
+    query.bindValue(":on_any_runway", flightCondition.onAnyRunway);
+    query.bindValue(":in_parking_state", flightCondition.inParkingState);
     query.bindValue(":in_clouds", flightCondition.inClouds);
     // No conversion to UTC
     query.bindValue(":start_local_sim_time", flightCondition.startLocalTime);
@@ -188,6 +197,9 @@ bool SQLiteFlightDao::get(std::int64_t id, Flight &flight) const noexcept
         const int titleIdx = record.indexOf("title");
         const int descriptionIdx = record.indexOf("description");
         const int surfaceTypeIdx = record.indexOf("surface_type");
+        const int onAnyRunwayIdx = record.indexOf("on_any_runway");
+        const int inParkingStateIdx = record.indexOf("in_parking_state");
+        const int surfaceConditionIdx = record.indexOf("surface_condition");
         const int groundAltitudeIdx = record.indexOf("ground_altitude");
         const int ambientTemperatureIdx = record.indexOf("ambient_temperature");
         const int totalAirTemperatureIdx = record.indexOf("total_air_temperature");
@@ -214,6 +226,9 @@ bool SQLiteFlightDao::get(std::int64_t id, Flight &flight) const noexcept
 
             FlightCondition flightCondition;
             flightCondition.surfaceType = static_cast<SimType::SurfaceType>(query.value(surfaceTypeIdx).toInt());
+            flightCondition.onAnyRunway = query.value(onAnyRunwayIdx).toBool();
+            flightCondition.inParkingState = query.value(inParkingStateIdx).toBool();
+            flightCondition.surfaceCondition = static_cast<SimType::SurfaceCondition>(query.value(surfaceConditionIdx).toInt());
             flightCondition.groundAltitude = query.value(groundAltitudeIdx).toFloat();
             flightCondition.ambientTemperature = query.value(ambientTemperatureIdx).toFloat();
             flightCondition.totalAirTemperature = query.value(totalAirTemperatureIdx).toFloat();